requirements.txt
python-telegram-bot==20.7
telethon==1.34.0
python-dotenv==1.0.0
flask==3.0.0
gunicorn==21.2.0
aiohttp==3.9.1
asyncio==3.4.3
Pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
ffmpeg-python==0.2.0
